[FIX] : User Document (MongoDb) extended FOSUserBundle/Entity/User (ORM)
